requirements.txt
Django
coverage
djangorestframework<4
py
pytest

twine==1.8.1
wcwidth==0.1.7
wheel